Erste Lösung, Test-Dependencies erweitert

+++ src/main/java/de/leetcode/problem1/Solution.java
@@ -2,16 +2,16 @@ package de.leetcode.problem1;
 
 public class Solution {
 
-    public int[] twoSum(int[] nums, int target) {
+    public long [] twoSum(long[] nums, int target) {
 
         for (int i = 0; i < nums.length; i++) {
             for (int j = i+1 ; j < nums.length; j++) {
                 if (nums[i] + nums[j] == target) {
-                    return new int[]{i, j};
+                    return new long[]{i, j};
                 }
             }
         }
         //Should not be reached
-        return new int[0];
+        return new long[0];
     }
 }
